Move to sensor factory
